Large-scale refactoring of files.
Move data handling into data_core.py, data_fetch.py and data_crunch.py.
data_core.py is intended as a high-level interface for main.py to reduce and simplify the functions needed to manage data.
data_fetch.py is responsible for getting, saving and loading data and making HTTP requests. It also has stub methods for deck-level and metagme-level data alongside the currently existing card-level data.
data_crunch.py contains various function that manipulate dataframes to isolate wanted information, or derive new information from exiting frames.
Multiple constants were moved from various files and added to settings.py
