+++ config/sets.config
@@ -1,4 +1,5 @@
 {	
+	"ADMINS" : ["ZacharyN#6195", "JasonILTG#4179", "Chord_O_Calls (Alex)#0385"]
 	"SETS" : ["MID"],
 	"FORMATS" : ["PremierDraft", "TradDraft", "QuickDraft"],
 	"DEFAULT_COLUMNS" : ["Color", "Rarity", "ALSA", "# GP", "GP WR", "# GIH", "GIH WR"],
